signal-processor/requirements.txt
kafka-python==2.0.2
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
asyncpg==0.29.0
pydantic==2.5.0
python-dateutil==2.8.2 